UI_PLAN Phase 3: HUD category tabs
The bar was every placeable building in one row. That works at twelve
and runs into the right-hand buttons somewhere past twenty — the second
capacity cliff UI_PLAN measured, and the one Phase 2's categories were
collected for. The bar is now 112px with a 28px tab strip: one tab per
BuildingCategory, each holding its own row of slots.

hud_view.c (SDL-free) owns layout, tabs, affordability and hit decoding;
ui.c is its drawer. The four separate hit-test entry points are gone —
ui_hit_test, ui_cog_hit_test, ui_demolish_hit_test, ui_world_hit_test
each recomputed geometry the drawer computed too, which is how a button
ends up a few pixels from where it appears. main.c now makes one
hud_hit() call against the list that was drawn.

Two decisions worth stating:

  - Greyed means MUTED, not disabled. An unaffordable building keeps its
    slot, its position and its clickability, because the build-confirm
    popup can still offer to pay in Gold — refusing the click would take
    away a real option. UI_W_MUTED exists for that distinction;
    UI_W_DISABLED stays for things that genuinely cannot be done. The
    test asserts slots do not move when affordability changes, since a
    slot that shuffles as Gold crosses a threshold is worse than one
    that greys.
  - The tab is sticky: it changes when a tab is clicked and at no other
    time. Asserted, because "helpfully" following the player's last
    build would be pleasant once and disorienting thereafter.

HUD metrics moved from ui.h into hud_view.h (which ui.h includes): they
had to leave a header that carries SDL, since the layout using them may
not. Every existing reader of HUD_HEIGHT is unchanged.

Verification: tests/test_hud.c drives the real layout and hit-test with
synthetic def tables of 12, 25 and 40 buildings, asserting the bar stays
on screen and slots never reach the right-hand cluster; plus tab
filtering, stickiness, the greyed-not-hidden rules, every widget
hit-testing back to itself, and the real def table showing each
placeable building in exactly one tab. 14 tests, 275 assertions, 0
failures; clean -Werror build; smoke and SDL-free checks pass.

+++ tests/test_defs.c
@@ -65,7 +65,7 @@ int main(void)
             printf("  FAIL: def[%d] (%s) has no category\n", i,
                    BUILDING_DEFS[i].name ? BUILDING_DEFS[i].name : "?");
             failures++;
-        } else if (BUILDING_DEFS[i].category >= BCAT_COUNT) {
+        } else if ((int)BUILDING_DEFS[i].category >= BCAT_COUNT) {
             printf("  FAIL: def[%d] has an out-of-range category\n", i);
             failures++;
         }
@@ -74,7 +74,7 @@ int main(void)
 
     for (i = 0; i < RES_COUNT; i++) {
         if (RESOURCE_CATEGORIES[i] == RCAT_NONE ||
-            RESOURCE_CATEGORIES[i] >= RCAT_COUNT) {
+            (int)RESOURCE_CATEGORIES[i] >= RCAT_COUNT) {
             printf("  FAIL: resource %d (%s) has no category\n", i,
                    RESOURCE_NAMES[i] ? RESOURCE_NAMES[i] : "?");
             failures++;
@@ -114,7 +114,7 @@ int main(void)
         for (c = 1; c < BCAT_COUNT; c++) {
             int n = 0, t;
             for (t = 0; t < BUILDING_TYPE_COUNT; t++)
-                if (BUILDING_DEFS[t].category == c &&
+                if ((int)BUILDING_DEFS[t].category == c &&
                     BUILDING_DEFS[t].hud_placeable) n++;
             if (n == 0) {
                 printf("  FAIL: category '%s' has no placeable buildings\n",
